Moved sendPasswordResetNotification to own trait
