+++ src/Models/Concerns/IsFilamentUser.php
@@ -2,9 +2,6 @@
 
 namespace Filament\Models\Concerns;
 
-use Illuminate\Auth\Notifications\ResetPassword as ResetPasswordNotification;
-use Illuminate\Support\Facades\URL;
-
 trait IsFilamentUser
 {
     public function canAccessFilament()
@@ -77,22 +74,5 @@ trait IsFilamentUser
         return $column !== null ?
             $this->{$column} :
             false;
-    }
-
-    public function sendPasswordResetNotification($token)
-    {
-        $notification = new ResetPasswordNotification($token);
-        $notification->createUrlUsing(function ($notifiable, $token) {
-            return URL::signedRoute(
-                'filament.auth.password.reset',
-                [
-                    'email' => $notifiable->getEmailForPasswordReset(),
-                    'token' => $token,
-                ],
-                now()->addMinutes(config('auth.passwords.filament_users.expire')),
-            );
-        });
-
-        $this->notify($notification);
     }
 }
