+++ src/Models/Concerns/SendsPasswordResetNotification.php
@@ -1,0 +1,26 @@
+<?php
+
+namespace Filament\Models\Concerns;
+
+use Illuminate\Auth\Notifications\ResetPassword as ResetPasswordNotification;
+use Illuminate\Support\Facades\URL;
+
+trait SendsPasswordResetNotification
+{
+    public function sendPasswordResetNotification($token)
+    {
+        $notification = new ResetPasswordNotification($token);
+        $notification->createUrlUsing(function ($notifiable, $token) {
+            return URL::signedRoute(
+                'filament.auth.password.reset',
+                [
+                    'email' => $notifiable->getEmailForPasswordReset(),
+                    'token' => $token,
+                ],
+                now()->addMinutes(config('auth.passwords.filament_users.expire')),
+            );
+        });
+
+        $this->notify($notification);
+    }
+}
